+++ src/components/upload/FileErrors/FileErrors.css
